Power button handler only removes valid timeout IDs
The function RemoveTimeoutIfSet in power_button_handler.cc calls
g_source_remove on a given timeout but does not check if the ID is valid
(greater than 0).  It manifested itself in the PowerDaemon unit tests.
glib has an assertion that complained when g_source_remove was passed an
ID of 0.

This patch adds a check for ID > 0.

BUG=chromium-os:8018
TEST=Ran unit tests and the assertion no longer appears.

Change-Id: Ibb03632c7b373adbe29dcd57b7d696d7dc1f3c4a
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/4002003