power_manager : Disable virtual terminal during suspend
Added functions to lock and unlock virtual terminal (VT) switching. The
lock function is called by the powerman daemon when going into suspend.
The unlock function is called by the dbus message handler when it
receives the signal from powerd_suspend indicating a resume.

The problem was that when going into suspend, the system switches to
VT10 before actually suspending. When resuming, it starts in VT10
before restoring the original VT. This is an unnecessary delay in the
suspend/resume cycle. By locking VT switching, this switch no longer
happens, and the suspend/resume time is reduced.

(brought from http://codereview.chromium.org/3595011)

This change selectively enables the virtual terminal
switching depending on use flags.
This is a short term fix for dogfood devices which have
failing touchpads when we don't console switch.

Change-Id: I71df16dc36ec2df1a642e5942dae628a0f197a76
Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chrome-os-partner:1400
TEST=(on mario) Close lid to suspend, open to resume,
there is no longer a VT10
screen (blank except for a cursor).
(on dogfood) Close lid to suspend, open to resume.
Ensure that there is a console switch to VT10,
and that mouse is functional

Review URL: http://codereview.chromium.org/4039005