power_manager : GetConsole before checking for lid closed.
GetConsole before querying for lid state. Lid state
query may result in the system suspending, which will fail
a CHECK because we have not gotten the console for vt
switch locking yet.

Change-Id: I931de659cd362e77e67ea4cbb9aa99e022ee0603
Signed-off-by: Benson Leung <bleung@chromium.org>
Signed-off-by: Todd Broch <tbroch@chromium.org>
Signed-off-by: Sameer Nanda <snanda@chromium.org>

BUG=chromium-os:8219
TEST=Boot the system, close the lid after 2 seconds.
Verify that the system goes to sleep, and that
upon resuming, the power  button works, indicating
that powerm has not crashed.

Review URL: http://codereview.chromium.org/4126003