power_manager : Flexible SetIdleOffset, No OOBE Screenoff
Make SetIdleOffset more flexible by being able to specify
the idle states on which to apply an offset. Allows for delaying
off state and suspend state, but not bringing the system back
up to Normal in the process.

This allows for keeping us in a state no lower than kIdleDim
while an OOBE update is in progress.

Change-Id: I71f79e87afb6d5b2989fa282f312d279af441a4e
Signed-off-by: Benson Leung <bleung@chromium.org>
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Signed-off-by: Nikita Kostylev <nkostylev@chromium.org>

BUG=chromium-os:7862
BUG=chromium-os:5348
TEST=rm /home/chronos/.oobe_completed
Verify that the system will dim screen, but not screen off
due to idle activity.
Create .oobe_completed in correct place, allow the system to
enter screenoff mode after 180s while on batt power. Plug in ac power.
Verify the screen does not come on.
With AC power attached, idle > 10 minutes. Unplug power.
Verify that system does not immediately suspend. Verify that it will
wait for 10 more minutes of idle on batt power before idle suspend.

Review URL: http://codereview.chromium.org/4175009