cromo: catch dbus-c++ exceptions in RegisterSuspendDelayCallback.
If pwoerd vanishes between us getting the NameOwnerChanged indicating its birth
and the NameOwnerChanged indicating its death, we'll try to send to a
nonexistent target on the bus, which dbus-c++ indicates with an exception. Catch
and ignore it - we'll see the queued NameOwnerChanged messages and do the right
thing.

BUG=chromium-os:9023
TEST=Adhoc
Inhibit powerd from starting, then add an unconditional call to PowerDaemonUp()
in CheckForPowerDaemon(). Without this change, cromo will abort; with it, you'll
get an error message in syslog and cromo will carry on.

Change-Id: I4733b9340466f9eb83b08afe4d40d4836b8f6cf1
X-Where: locus chromeos-dogfood vcsid-fix
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Signed-off-by: Jason Glasgow <jglasgow@chromium.org>

Review URL: http://codereview.chromium.org/4897001