gobi-cromo-plugin: Add internal modem reset.
Previously, if the modem got into an unrecoverable state, we'd just abort and
let the upstart script reset the modem for us; this is kind of messy and causes
fake cromo crashes, because we have to exit abnormally to trigger that part of
the script. Instead, we'll try invoking a setuid helper program to do a reset
for us, and then try to open the modem again.

BUG=chromium-os:9063
TEST=Adhoc
Hack gobi_sdk_wrapper.cc a bit - add something like:
	static int fail = 0;
	if (!fail++)
		return 12;
to the start of QCWWANConnect() to simulate a failure being recovered by the
reset. The modem should still function normally.
--
Change-Id: I8dbf03b10ce3e38b2c9683281c9b348ead890669
X-Where: locus chromeos-dogfood gobi-reset
Signed-off-by: Elly Jones <ellyjones@google.com>
Signed-off-by: Jason Glasgow <jglasgow@google.com>

Review URL: https://chromereview.prom.corp.google.com/2093009