Added udev rule and setup scripts for light sensor
There was a problem with the initialization of the light sensor.  It was
being initialized with upstart.  However, since the sensor is on the i2c
bus, it should not be initialized until the i2c is set up.  The upstart
script was being run before i2c was ready.  The result was that the
light sensor calibration value was not set properly, and the screen did
not dim noticeably when the sensor was covered.

This patch replaces the upstart setup with a udev rule that calls the
setup scripts when i2c_i801 is installed.  The setup scripts install
tsl2563 and set the light sensor calibration value.

You can see the udev events using "udevadm monitor" in one console while
installing/removing i2c_i801 module in another.

BUG=chrome-os-partner:591
TEST=Manually install the scripts to /etc/udev and rules file to
/etc/udev/rules.d, since ebuild changes are separate from this patch.
Restart system and test light sensor.  Screen should dim visibly.

Signed-off-by: Simon Que <sque@chromium.org>
Change-Id: I41e338d8f6a8cd9db6e9467221ea1ebd1a193add

Review URL: http://codereview.chromium.org/5201001