Shut down instead of suspend at 1% battery
Shut down instead of suspend at 1% battery to dissuade
users from draining the system all the way down to 0%
and uncleanly shutting down.

This allows us to avoid crashing chrome or failing to
disconnect from the wireless carrier in a low battery
condition.

Previously, the system would enter suspend state, which
was easy to wake from due to a keypress, lid or trackpad
move.

Also, by shutting down, the system will continue to shut
down after booting to the login screen as powerd
continues to detect that we are below the 1% watermark.

This is the behavior that we want to encourage the user
to find AC power ASAP.

Change-Id: I8a017576248f44de9341acdea7d8aae5d55ccfcd
Signed-off-by: Benson Leung <bleung@chromium.org>
Signed-off-by: Simon Que <sque@chromium.org>

BUG=chromium-os:8979
TEST=Drain the battery down to 1% (or change
low_battery_suspend_percentage to a higher value).
Ensure the system shuts down completely after getting below
1% on batt. Ensure that booting the system afterward without
plugging back into AC power results in shutdown at the
login screen.

Review URL: http://codereview.chromium.org/5132008