Record uptime info during logout
This is part of effort to investigate and fix slowness in logout.
Note that "bootstat foo" will create /tmp/uptime-foo and /tmp/disk-foo.

Those files are copied to /home/chronos/prev_stats as well
as /tmp will be deleted upown reboot of the machine.

The original patch was reverted as it broke the official build.
"cp" failed if some files were not present in /tmp. It seems files like
/tmp/login-times-sent are not present in some cases. In this version,
change the cp command line to ignore errors.

BUG=chromium-os:6847
TEST=confirmed that /tmp/uptime-* files were created as expected. Confirmed that it worked on the official build on VM.

Review URL: http://codereview.chromium.org/5077001

Change-Id: I9666dddb6f060b63a5514c0d699247f70072a17d