powerd supports systems without backlight
Updated powerd to be able to run on systems that don't have a backlight.
Instead of fatally exiting when backlight initialization fails, it
should proceed gracefully.  The backlight module has been given an
initialized flag that is set when it is initialized successfully.  If
not, its functions are "stubbed out" by an early return.  This allows
the powerd daemon and BacklightController to call backlight functions
even when backlight doesn't exist.  This removes the need to add lots of
checks to Daemon and BacklightController.

BUG=chromium-os:9631
TEST=Run powerd on a system without backlight, like qemu VM.  It should
start successfully.

Change-Id: Ie6d5a71784e325a6d75e4db7c85ea655ee31245f
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/5350002