userfeedback: remove wpa_supplicant log
The /var/log/hostap.log file hasn't been used since we turned off
wpa_supplicant logging.  This used to not be an issue but recent
mods to userfeedback cause a tab to be recreated with <no value>.
Fix this by removing the entry from the .lst file.

BUG=none
TEST=gmerge to device; run chrome:system and verify the wpa_supplicant tab is gone

Review URL: http://codereview.chromium.org/4165006

Change-Id: Ie8f63b58d6752378b51b1d21a96da029d7d35d27