power_manager: Show animation for shutdown/restart requests.
This makes the power manager send messages to the window
manager telling it to display the zoom-out animation
when we get a request from Chrome to shutdown or restart
instead of just when the user is holding the power button.

This covers Chrome-initiated shutdowns such as those
triggered by the login screen's "Shut down" button or
restarts after updates are applied.  It also handles
non-Chrome shutdowns caused by low power or suspending while
not logged in.

Change-Id: Ibc3b3ed20cffc687530f0dd9bb2ef21edaff23d7

BUG=chromium-os:9848
TEST=clicked the shut down and restart buttons and checked that the animation was played.  also shut down by holding the power button to make sure that it didn't regress

Review URL: http://codereview.chromium.org/5519002