Restructure FilesystemCopierAction to use ping-pong buffers.
This would allow some parallel processing -- calculating hashes in parallel with
read in parallel with write. Performance benefit at this stage seems
marginal. The major reason for the change is to reuse the FilesystemCopierAction
for applied update verification -- hopefully speeding it up and making AU
responsive to D-Bus signals during that stage. That would be a separate CL
though.

BUG=9140
TEST=unit tests, tested update on the device

Change-Id: I9d07ebbf4d536362c0a955addcaa1d48d5e9b115

Review URL: http://codereview.chromium.org/5603001