powerm: input: save lid_fd_ whenever a lid is found
When "Watched event also has a lid!" occurs
when there is a single input event that contains both
power button and lid, we skip saving the lid_fd_,
and therefore the QueryLidState does not work.

Always save the lid fd.

Change-Id: Ibaad4511ec6162d999055a7061d87c649ef97ee6
Signed-off-by: Benson Leung <bleung@chromium.org>
Signed-off-by: Sameer Nanda <snanda@chromium.org>

BUG=chrome-os-partner:1788
TEST=boot the system with lid closed.
As it is booting, it should enter suspend.

Review URL: http://codereview.chromium.org/5532002