init: Run iptables and vt-switching even if UI fails.
This adds an "or stopped failsafe" to the start conditions
for these jobs (and renames the "sleep" job to "failsafe"),
so they'll eventually get run even if the UI fails to start
correctly.  I also updated a bunch of outdated mailing list
addresses in these files.

Change-Id: I009e991e1ea216db83b0d7e585025131c6c585de

BUG=chromium-os:10349
TEST=built image with updated init scripts and booted system.  added a "sleep 60" to the end of boot-complete.conf and checked that i'm able to ssh to the machine just 30 seconds after login-prompt-visible, when the failsafe job fires

Review URL: http://codereview.chromium.org/5871003