AU: Function to trigger crash reporter.
This CL introduces a new utility function that fork()s and has the
child call abort(), thus triggering the crash reporter. The actual
work is done in a main loop callback to ensure that the stack trace is
the same for all calls to this function.

TEST=unittest (minimal), tested on device - saw crash reporter pick up the crash.
BUG=10739

Review URL: http://codereview.chromium.org/6114002

Change-Id: If3c3c43644173f2307b8498d40881c3b38b938e5