AU: Reduce instances of signalling the crash reporter.
We recently started notifying the crash reporter whenever we attempted
to report an error code to the update server. We now throttle that
back a little bit: we only notify for official builds when we were
successful in reporting the error to the update server.

TEST=unittest; tested on device
BUG=10805

Review URL: http://codereview.chromium.org/6154003

Change-Id: Iaff1d5a0b600663796e63ad308f408c57d6444da