[login_manager] Prevent non-children from calling RestartJob
We were allowing any process to restart chrome, with an arbitrary set
of command-line options.  Disallow this by checking the pid of anyone who
attempts to RestartJob and only allowing the call to succeed if it's issued
by a process managed by session_manager.

BUG=10877
TEST=suite_Smoke; install on device and use dbus-send to attempt to restartjob, and see that it won't work

Change-Id: Ia6586c80c2325f557566d162a15da444dfed72ce

Review URL: http://codereview.chromium.org/6239002