AU: Add options to delta_generator to hash and sign generated payloads.
To generate a SHA-256 hash for signing an unsigned payload use:

./delta_generator -in_file <unsigned_payload> \
  -out_hash_file <hash_for_signing> -signature_size <size>

To add a signature to an unsigned payload use:

./delta_generator -in_file <unsigned_payload> \
  -signature_file <signature> -out_file <signed_payload>

The size of the signature file must be exactly the same as the signature size
supplied while hashing the payload.

Also removed the -apply_delta option because it's replaced by the new -in_file
option -- if -in_file is specified with no other hashing/signing option, the
generator runs in delta application mode.

BUG=chromium-os:10872
TEST=manually ran delta_generator on payloads

Change-Id: Ie918239a4db607153932cc1907127790cf3179d3

Review URL: http://codereview.chromium.org/6305002