Added GSM support by adding a GobiGsmModem subclass. This work is not yet complete, as some functions still need to be filled in (e.g., SendPin), but basic operation is correct. Connections can be established, and signal strength, network technology, and roaming state are communicated via D-Bus.
BUG=chromium-os:4667
TEST=Set carrier to T-Mobile (a GSM carrier) and, running with
modified versions of the Chrome and the flimflam modemmgr plugin,
verified that connections can be made, and that web browsing works.
Observed that signal strength and roaming status were displayed
correctly. Observed that when the modem was reporting that EDGE
was in use, the signal strength icon was overlayed with an "E"
badge, and when UMTS or HSPA were reported, the "3G" badge was
overlayed.

Change-Id: I2255f851e78b965c580b7c6d771bdbd59cd112a7

Review URL: https://chromereviews.googleplex.com/2357012