Copy PRL file to /tmp before issuing the command to cromo to update the PRL.
The "modem prl-update" command was failing for downloaded PRL files.
This was because the Downloads directory is owned by user "chronos",
and has permissions 700. cromo runs as user "cromo", and therefore
is not able to access the Downloads directory. So even though crosh,
which runs as "chronos", was able to read the file, cromo could not.
The solution is to have the modem script first copy the PRL file into
a public directory, /tmp.

BUG=chromium-os:9172
TEST=Before and after test, in which I tried the modem update-prl
command, giving it a PRL pathname in /home/chronos/user/Download,
and observed that it failed because cromo got Permission Denied
while trying to open the file. After the change was in place, tried
the same command and it worked, and afterwards, a copy of the PRL
file could be found in /tmp.

Change-Id: I0e58f64203109e9376a9cdcf36679b55c17b9f2d

Review URL: https://chromereviews.googleplex.com/2468012