Fixed a dbus assertion failure in the Gobi cromo plugin.
There was a bug in which the property map returned by GetStatus could
have two values for the "activation_state" property. That caused an
assertion failure in dbus, because it expected to write out a <string,
DBus::Variant> for each property, not <string, DBus::Variant, DBus::Variant>.
This problem only showed itself if the modem was in CDMA mode, and was
in the unactivated or partially activated state. The problem emerged
in the refactoring of GobiModem into GobiCdmaModem and GobiGsmModem,
which changed the order in which various calls were made.

BUG=chromium-os:11043
TEST=On a machine in the partially activated state with the carrier
set to Verizon Wireless, the crash would occur every time cromo
started up. After the change, no crash, and things work as expected.

Change-Id: I24eb076cde0e99f7242cc9278f7f4ab25baa354e

Review URL: https://chromereviews.googleplex.com/2476012