Smooth transitions in power manager
This patch implements gradual brightness change support for power manager.
This allows the gradual changes to be handled independently of the
kernel backlight driver, which can be platform-dependent.

To achieve the gradual smoothing, glib events are used.  When a new
brightness change is issued before the old transition has finished, it
obsoletes the steps that remain in the old transition, and proceed with
the new transition.

BUG=chromium-os:7615
TEST=With a kernel backlight driver that does not support gradual
transitions, adjust the brightness.  The transitions should appear
smooth.  This should work with both single key presses and holding down
the keys.

Signed-off-by: Simon Que <sque@chromium.org>
Change-Id: I8997e734ff499febfd4aac68f8aec3672c9437d0

Review URL: http://codereview.chromium.org/6327008