backlight-tool supports smooth transitions
Because of the addition of smooth brightness transitions in power
manager, the behavior of Backlight::SetBrightness has changed.  Instead
of directly updating the backlight brightness, it scheduled glib events
in the future to gradually change it over a period of time.

backlight-tool is dependent on SetBrightness, but it did not have a glib
loop, so it was broken after SetBrightness was updated.

This patch fixes that problem by updating backlight-tool.  It now has a
glib main loop to run the events in a gradual transition.

BUG=chromium-os:11306
TEST=Use 'backlight-tool --set_brightness <value>' to change the
brightness around.  There should be an observable gradual transition.

Change-Id: If70bb954033d5ef7144e377c2a9b46bb2126d9bb
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6397001