Disable light sensor adjustment if backlight is off
The ambient light sensor should not adjust the brightness if it is set
all the way to zero.  The backlight, when turned off, should stay off.
This patch implements that condition.

BUG=chromium-os:7615
TEST=Set brightness to zero and shine a light into the light sensor.
The backlight should remain off.

Change-Id: I2ff56a1a1bec6497cb86ea9fb0a6965abfa11895
Signed-off-by: Simon Que <sque@chromium.org>

Debug: reduced dim and off times

Change-Id: I3c7938a6c624157b533ea2f04e94081bb6cb8559

backlight-tool supports smooth transitions

Because of the addition of smooth brightness transitions in power
manager, the behavior of Backlight::SetBrightness has changed.  Instead
of directly updating the backlight brightness, it scheduled glib events
in the future to gradually change it over a period of time.

backlight-tool is dependent on SetBrightness, but it did not have a glib
loop, so it was broken after SetBrightness was updated.

This patch fixes that problem by updating backlight-tool.  It now has a
glib main loop to run the events in a gradual transition.

BUG=chromium-os:11306
TEST=Use 'backlight-tool --set_brightness <value>' to change the
brightness around.  There should be an observable gradual transition.

Change-Id: If70bb954033d5ef7144e377c2a9b46bb2126d9bb
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6326015