Do not start with low brightness
If the brightness is set to a very low value, the user runs the risk of
not being able to see what's on screen after a shutdown and restart.
This patch adjusts the stored brightness offset so that the resulting
brightness will be nonzero.  It also adjusts the offsets when reading
from prefs.  Additionally, ALS values are stored as part of the power
prefs now.

BUG=chromium-os:11458
TEST=Set brightness to zero.  Stop and restart powerd.  The backlight
should be on, albeit dim.  Once again, set brightness to zero and
stop powerd.  In /var/lib/power_manager, delete als_brightness_level,
write "0" to the current brightness offset, and start powerd.  The
backlight should turn on.

Change-Id: I3b149f6de9ab45ea02466649315e2b6bd09b8d0e
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6312050