Avoid being stuck at low brightness due to ALS adjustments
When adjusting backlight due to a change in ALS value, we don't want it
to set the brightness to zero.  The old solution was to modify the
stored ALS reading so ALS + brightness offset > 0.  This resulted in
the local value being out of sync with future ALS readings, causing
improper brightness adjustments.

This patch fixes the problem by tweaking the brightness offset instead
of the stored ALS level to make sure brightness > 0.

BUG=chromium-os:11797
TEST=1. Shine a bright light at the ambient light sensor.
2. press the Brightness Up key until backlight is fully bright
3. press the Brightness Down key until it is the last setting before
backlight is disabled
4. Remove bright light from als.
5. Press the brightness key up key.
6. Expected output: Brightness adjusted up one notch. ALS is stable,
and doesn't cause fluctuating brightness.

Change-Id: Iba6a0456ccde1c4163ff4ae42466124ce1f7832e
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6489008