More meaningful log messages in SetBrightness
This patch improves the traces in Backlight::SetBrightness.  The
new target brightness is only set if the requested brightness does not
match the current brightness.  The new traces print out the current
brightness and current target brightness, in addition to the requested
brightness which is already printed.

BUG=chromium-os:11998
TEST=Run unit tests, make sure they pass.

Change-Id: I10767614ff27c6e03e132956a2a4ed2b0d645df0
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6480055