Avoid zero brightness during plug in/out transition
powerd keeps two brightness offsets, one for plugged and one for
unplugged.  Because they are used with the ALS value, it is still
possible to drop brightness to zero by plugging in or unplugging the
power, because powerd doesn't check for zero brightness during plug
events.  It only makes that check during ALS readings.

This patch fixes that problem by checking for zero brightness and
adjusting brightness offset so that brightness > 0 when switching
between plugged and unplugged.

BUG=chromium-os:12142
TEST=Go to battery power.  Shine light in ALS and turn down brightness
to just above zero.  Plug in AC power and remove light.  Unplug power.
The brightness should not drop to zero.

Change-Id: I089d227fd262d5e10484a38574bba5c7098ce3d8
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6529029