Avoid double brightness changes at low brightness
There seems te a double WriteBrightness when setting from 1 to 0, in
rapid succession.  This results in SetBrightness being called twice with
the same sequence.  This was not caught by Benson's earlier change to
reduce redundancies because the backlight brightness has not had time to
change before the second write/set is called.

BUG=chromium-os:11728
TEST=Do 'tail -f /var/log/power_manager/powerd.LATEST', set brightness
to full and then down to 1.  Set it down once more and break out of
'tail -f' and inspect log.  When setting brightness to 0, there should
not be duplicate brightness change steps.

Change-Id: I7a5630592e120e46900b2ed2797af76d9d16bdab
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6556001