[login_manager] Add DevicePolicy, which will handle persisting/loading policies.
We're working towards replacing the PrefStore with a DevicePolicy,
which will really be a serialized protobuffer and an associated
signature.  Clients will get/set this policy as a blob, so the API is
much simpler.  We still put the policy and the signature into a JSON
file because it's a convenient way to bind them together to make sure
they are written to disk atomically.

BUG=12670
TEST=unit tests

Change-Id: I3e4796b9ab9a4631edc761f8aac7fdc62e66215d

Review URL: http://codereview.chromium.org/6659011