Added uninitialized powerd state and improved brightness transition logic
Added an UNINITIALIZED power state to the backlight controller, which is
used as the starting state.  This way, backlight_controller will
initialize ALS correctly.  Before, ALS was not being initialized until
there was a power state transition.  Now, it will be initialized when
backlight_controller state is set to ACTIVE_ON during initialization
(but after the constructor has been called), because that is considered
a transition from UNINITIALIZED.

Additionally, the brightness transitions logic is improved.  It will
properly handle the sort of scenario where the existing target brightness
is 0 and the current brightness is 3, and a call to set brightness to 3
is made.  The old code would disregard the call because the current
brightness was equal to the new brightness.

However, the brightness would then be set to 0 by the existing transition,
when the latest request was for a brightness of 3.  The solution is to set
the new target brightness before checking target brightness against current
brightness.

BUG=chromium-os:12790
TEST=Boot up and get to the login screen.  Cover the ALS or shine a
light into it.  The brightness should change.
Also, set brightness to just above zero.  Run "stop powerd" and "start powerd"
to restart powerd.  Make sure brightness is not at zero when powerd starts.

Change-Id: I1b7ea1ecf6ee0161a8ab06c9f038923cd058fe5f
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6649004