Do not send activate state change on success.  The modem is going to reset itself anyway.
If a an activation state change messages is sent, flimflam and the ui respond by immediately sending a connect request.  That connect request races with the modem reset, and when it arrives first leaves the modem in an indeterminate state.  One from which it will not recover unless the modem is powered off.

This change requires http://codereview.chromium.org/6709018/, because otherwise flimflam holds a reference to the hold (reset) modem and never deletes the old stale object.

BUG=chromium-os:12913
TEST=Run through activation with Verizon twice

Review URL: https://chromereviews.googleplex.com/2677022

Change-Id: I9308c3db2d36708276ca7a4e04662a13226845b6